LPS-99564 Moves code to EditFragmentEntryDisplayContext
